level 2 warnings nailed

+++ r_bsp.cpp
@@ -276,7 +276,7 @@ void R_AddLine (seg_t*	line)
 	// Totally off the left edge?
 	if (tspan >= span)
 	    return;	
-	angle2 = -clipangle;
+	angle2 = -static_cast<int>(clipangle);
     }
     
     // The seg is in the view range,
@@ -428,7 +428,7 @@ bool R_CheckBBox (fixed_t*	bspcoord)
 	if (tspan >= span)
 	    return false;
 	
-	angle2 = -clipangle;
+	angle2 = -static_cast<int>(clipangle);
     }
